Added col checks

+++ app/database/migrations/2019_02_09_162734_add_marker_id_to_project_table.php
@@ -15,11 +15,13 @@ class AddMarkerIdToProjectTable extends Migration {
 			foreach(get_education_levels() as $key => $level) {
 				$projectTableName = $department.'_projects_'.$level['shortName'];
 
-				// Add the marker id to the project table
-				Schema::table($projectTableName, function (Blueprint $table) use ($department, $level){
-					$table->uuid('marker_id')->nullable(true);
-					$table->foreign('marker_id')->references('id')->on($department.'_supervisors')->onDelete('SET NULL');
-				});
+				if(!Schema::hasColumn($projectTableName, 'marker_id')){
+					// Add the marker id to the project table
+					Schema::table($projectTableName, function (Blueprint $table) use ($department, $level){
+						$table->uuid('marker_id')->nullable(true);
+						$table->foreign('marker_id')->references('id')->on($department.'_supervisors')->onDelete('SET NULL');
+					});
+				}
 			}
 		}
 	}
