+++ app/database/migrations/2019_02_09_162736_remove_marker_id_from_student_table.php
@@ -14,10 +14,13 @@ class RemoveMarkerIdFromStudentTable extends Migration {
 		foreach(get_departments() as $key => $department) {
 			foreach(get_education_levels() as $key => $level) {
 				$studentTableName = $department.'_students_'.$level['shortName'];
-				// Remove the marker id from the student table
-				Schema::table($studentTableName, function (Blueprint $table) use ($department, $level){
-					$table->dropColumn('marker_id');
-				});
+
+				if(Schema::hasColumn($studentTableName, 'marker_id')){
+					// Remove the marker id from the student table
+					Schema::table($studentTableName, function (Blueprint $table) use ($department, $level){
+						$table->dropColumn('marker_id');
+					});
+				}
 			}
 		}
 	}
